MAJOR NAMING CONVENTION CHANGE (please read)
Before the explanation, I’d like to apologize for such major change this
late in development. Even though the mod has not been released yet it is
available on GitHub for some time, so it’s possible that some people may
already write some BG1 related code with EET in mind. If you are one of
this people than please send me your code/files and I will update them
myself, so you don’t have to waste time due to this change. Here are the
changes:
1. BG:EE area resources prefix changed from FH to BG.
I've found a list of all new BP-BGT Worldmap area names that modders
use:
https://kerzenburg.baldurs-gate.eu/showthread.php?t=35175
It looks like none of mods used BG prefix for area names, so after some
serious thought I decided to adopt it. The reason behind this change is
to make it easier to remember for normal players who may use
walkthroughs. This way the only thing they need to know is acronym of
the game ;) Second thing is that FH looks similar to FW (area prefix
known from Tutu), which may lead to confusion and mistakes in code (in
fact up until recently there was a bug in EET worldmap tables (which has
been ported from Tutu), where I didn't noticed this name difference
myself).

2. Spells from BG:EE that conflicts with BG2:EE has now BG prefix that
replaces SP vanilla one, instead of "_" at the end. The reason behind
this is a problem with Opcode #172 that can’t remove spells that have 8
characters:
http://gibberlings3.net/forums/index.php?showtopic=26474