Remove manual steps in favor of PCU duplicate.
DRGRBRHT.SPL is referred in the SoD assets both as fullcaps and
mixed-caps DrGrBrHt in the BCS scripts. In the previous PR, I attempted
to resolve this via a single PCU entry and then doing the other steps
manually, but that led to some changes being attempted to be installed
twice.

This PR drops all the manual steps introduced in the prior PR in favor
of duplicating the PCU entry to catch both spelling variations.