Update README.md
Copy-edited the README text for grammar and punctuation. It's a very little thing, but this is the first text that potential contributors will see when they view this project on GitHub and it will help set a good first impression. I did not otherwise add or alter the content of what had already been written here.

I don't currently have the modding ability to contribute to this project, but I want to help you guys out in this small way.